Split advancedInvite into cohesive modules
The inviter authoring bench's pure model had grown into one ~1570-line
file spanning three unrelated concerns. Split it along those concerns
into sibling modules under apps/web/src/psi, keeping advancedInvite.ts
as a thin barrel that re-exports the same public surface so all eight
importers and the existing tests keep resolving @psi/advancedInvite
unchanged.

The new boundaries:

advancedInviteTypes.ts holds the shared draft/seed types and the
OutputDirection <-> Output mapping pair, a leaf the other three build on.

advancedInviteTerms.ts holds the draft-to-LinkageTerms mapping and the
inviter's exchange-spec assembly. The gated-setting clamp (psi-c,
deduplicate, per-element fuzzy) lives here in buildAdvancedTerms.

advancedInviteDraft.ts holds the draft-editing operations: seeding,
metadata and standardization reconciliation, expert key/element
authoring, and reconstructing a draft from an imported terms document.

advancedInviteValidation.ts holds the Generate gate and the
import-refusal messages, including the psi-c import refusal.

The move is behavior-preserving: draft transformations, validation
outcomes, and terms mapping are byte-identical. The psi-c clamp, the
psi-c import refusal, and the terms-consistency guards are carried over
verbatim as layered mint-time protections, not removed as redundant.

Each split unit gains focused unit tests for the contract the pre-split
suite only exercised through the whole module: the OutputDirection
inverse and its safe "both" default, producibleFieldNames, the
keep-keys metadata reconcile, and the gating guards pinned to their new
module homes.
